blueprinthub/pyproject.toml
[tool.poetry]
name = "blueprinthub"
version = "0.1.0"
description = "A scalable CLI for templated project scaffolding and GitHub imports"
authors = ["Srinivas Sarkar <srinivassarkar07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.9.0"
jinja2 = "^3.1.2"
gitpython = "^3.1.40"
questionary = "^2.0.1"
pyyaml = "^6.0.1"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
mypy = "^1.8.0"
pytest-cov = "^4.1.0"
pre-commit = "^3.6.0"
black = "^24.2.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"